+++ projects/ngx-forms/src/lib/ngx-forms-ui/components/form-card/form-card.component.ts
@@ -1,11 +1,12 @@
-import {Component, Input, ViewEncapsulation} from '@angular/core';
+import {ChangeDetectionStrategy, Component, Input, ViewEncapsulation} from '@angular/core';
 
 
 @Component({
   selector: 'form-card',
   templateUrl: './form-card.component.html',
   styleUrls: ['./form-card.component.scss'],
-  encapsulation: ViewEncapsulation.None
+  encapsulation: ViewEncapsulation.None,
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class FormCardComponent {