+++ projects/ngx-forms/src/lib/ngx-forms-ui/components/form-flow-wrapper/form-flow-wrapper.component.ts
@@ -1,10 +1,11 @@
-import {Component, ViewEncapsulation} from '@angular/core';
+import {ChangeDetectionStrategy, Component, ViewEncapsulation} from '@angular/core';
 
 @Component({
   selector: 'form-flow-wrapper',
   templateUrl: './form-flow-wrapper.component.html',
   styleUrls: ['./form-flow-wrapper.component.scss'],
-  encapsulation: ViewEncapsulation.None
+  encapsulation: ViewEncapsulation.None,
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class FormFlowWrapperComponent {