+++ projects/ngx-material/src/lib/ngx-dialog/components/popup-controller/popup-controller.component.ts
@@ -1,10 +1,11 @@
-import {Component} from '@angular/core';
+import {ChangeDetectionStrategy, Component} from '@angular/core';
 import {PopupService} from "../../services/popup.service";
 
 @Component({
   selector: 'ngx-popup-controller',
   templateUrl: './popup-controller.component.html',
-  styleUrls: ['./popup-controller.component.scss']
+  styleUrls: ['./popup-controller.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class PopupControllerComponent {