+++ projects/ngx-service-worker/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @consensus-labs/ngx-service-worker
 
+## 1.0.6
+
+### Patch Changes
+
+- Update change detection to OnPush
+- Updated dependencies
+  - @consensus-labs/ngx-material@1.0.10
+  - @consensus-labs/ngx-tools@1.0.13
+
 ## 1.0.5
 
 ### Patch Changes