+++ projects/ngx-tools/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @consensus-labs/ngx-tools
 
+## 1.0.13
+
+### Patch Changes
+
+- Update change detection to OnPush
+- Add TimeAgo pipe
+
 ## 1.0.12
 
 ### Patch Changes